new-backend/requirement.txt
# FastAPI and Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
pillow>=10.0.0
pytesseract>=0.3.10

# Data Validation
pydantic>=2.6.0

# Environment Management
python-dotenv>=1.0.0

# LLM Integration (Optional but recommended)
google-generativeai>=0.3.0

# Logging and Monitoring (Optional)
structlog>=24.0.0